Revert "GEODE-6612: move attributes that can not be configured in v2 api out of RegionConfig (#3425)"
This reverts commit 2402cac0

* document what's supported in management v2 api.
