+++ geode-core/src/distributedTest/java/org/apache/geode/distributed/internal/CacheConfigDAODUnitTest.java
@@ -22,7 +22,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.Rule;
 import org.junit.Test;
 
-import org.apache.geode.cache.configuration.BasicRegionConfig;
+import org.apache.geode.cache.configuration.RegionConfig;
 import org.apache.geode.cache.configuration.RegionType;
 import org.apache.geode.distributed.ConfigurationPersistenceService;
 import org.apache.geode.test.dunit.rules.ClusterStartupRule;
@@ -44,10 +44,10 @@ public class CacheConfigDAODUnitTest {
       ConfigurationPersistenceService ccService =
           ClusterStartupRule.getLocator().getConfigurationPersistenceService();
       ccService.updateCacheConfig("cluster", cc -> {
-        BasicRegionConfig regionConfig = new BasicRegionConfig();
+        RegionConfig regionConfig = new RegionConfig();
         regionConfig.setName("regionB");
         regionConfig.setType(RegionType.REPLICATE);
-        cc.addRegion(regionConfig);
+        cc.getRegions().add(regionConfig);
         return cc;
       });
     });
