+++ geode-core/src/main/java/org/apache/geode/management/internal/configuration/validators/RegionConfigValidator.java
@@ -15,7 +15,6 @@
 
 package org.apache.geode.management.internal.configuration.validators;
 
-import org.apache.geode.cache.configuration.BasicRegionConfig;
 import org.apache.geode.cache.configuration.CacheConfig;
 import org.apache.geode.cache.configuration.CacheElement;
 import org.apache.geode.cache.configuration.RegionConfig;
@@ -24,7 +23,7 @@ import org.apache.geode.internal.cache.InternalCache;
 import org.apache.geode.internal.cache.RegionNameValidation;
 import org.apache.geode.security.ResourcePermission;
 
-public class RegionConfigValidator implements ConfigurationValidator<BasicRegionConfig> {
+public class RegionConfigValidator implements ConfigurationValidator<RegionConfig> {
   private InternalCache cache;
 
   public RegionConfigValidator(InternalCache cache) {
@@ -32,7 +31,7 @@ public class RegionConfigValidator implements ConfigurationValidator<BasicRegion
   }
 
   @Override
-  public void validate(BasicRegionConfig config)
+  public void validate(RegionConfig config)
       throws IllegalArgumentException {
     if (config instanceof RegionConfig) {
       throw new IllegalArgumentException("Use BasicRegionConfig to configure your region.");
@@ -75,7 +74,7 @@ public class RegionConfigValidator implements ConfigurationValidator<BasicRegion
   }
 
   @Override
-  public boolean exists(BasicRegionConfig config, CacheConfig existing) {
+  public boolean exists(RegionConfig config, CacheConfig existing) {
     return CacheElement.exists(existing.getRegions(), config.getId());
   }
 }
