+++ geode-core/src/test/java/org/apache/geode/cache/configuration/RegionConfigTest.java
@@ -68,8 +68,7 @@ public class RegionConfigTest {
       RegionConfig config = new RegionConfig();
       config.setType(shortcut.name());
       config.setName(shortcut.name());
-      RegionConfig masterRegion =
-          CacheElement.findElement(master.getRegions(), shortcut.name());
+      RegionConfig masterRegion = CacheElement.findElement(master.getRegions(), shortcut.name());
       assertThat(config).isEqualToComparingFieldByFieldRecursively(masterRegion);
     }
   }
